+++ saleor/webhook/payloads.py
@@ -1270,8 +1270,12 @@ def generate_checkout_payload_for_tax_calculation(
         user_public_metadata = user.metadata
 
     # Prepare discount data
-    is_shipping_voucher = (
-        checkout_info.voucher.type == VoucherType.SHIPPING
+    # total_amount include the specific product and apply once per order discounts,
+    # so we need to attach only entire order discount here with once per order flag
+    # set to False
+    discount_not_included = (
+        checkout_info.voucher.type == VoucherType.ENTIRE_ORDER
+        and checkout_info.voucher.apply_once_per_order is False
         if checkout_info.voucher
         else False
     )
@@ -1279,7 +1283,7 @@ def generate_checkout_payload_for_tax_calculation(
     discount_name = checkout.discount_name
     discounts = (
         [{"name": discount_name, "amount": discount_amount}]
-        if discount_amount and not is_shipping_voucher
+        if discount_amount and discount_not_included
         else []
     )
 
@@ -1291,6 +1295,11 @@ def generate_checkout_payload_for_tax_calculation(
     shipping_method_amount = quantize_price(
         base_calculations.base_checkout_delivery_price(checkout_info, lines).amount,
         checkout.currency,
+    )
+    is_shipping_voucher = (
+        checkout_info.voucher.type == VoucherType.SHIPPING
+        if checkout_info.voucher
+        else False
     )
     if is_shipping_voucher:
         shipping_method_amount = max(